@@ -18,12 +18,12 @@ declare module "path" {
 //// [mainJs.js]
 import {} from "./a";
 import("./a");
-const _ = require("./a"); // No resolution
+const _ = require("./a");
 _.a; // any
 
 //// [main.ts]
 import {} from "./a";
-import _ = require("./a"); // Error
+import _ = require("./a"); // Error in esnext
 export = {}; // Error
 export {};
 
@@ -35,8 +35,7 @@ export const a = "a";
 export var a = "a";
 //// [mainJs.js]
 import("./a");
-var _ = require("./a"); // No resolution
+var _ = require("./a");
 _.a; // any
-export {};
 //// [main.js]
-export {};
+module.exports = {};
