@@ -1,5 +1,7 @@
 import fs from "fs";
-import glob from "glob";
+import {
+    glob,
+} from "glob";
 import path from "path";
 import url from "url";
 
