@@ -11,6 +11,7 @@ import {
     ExportAssignment,
     ExportDeclaration,
     Expression,
+    ExpressionStatement,
     GeneratedIdentifierFlags,
     getEmitFlags,
     getEmitModuleKind,
@@ -88,7 +89,7 @@ export function transformECMAScriptModule(context: TransformationContext): (x: S
                     setTextRange(factory.createNodeArray(insertStatementsAfterCustomPrologue(result.statements.slice(), importRequireStatements)), result.statements),
                 );
             }
-            if (!isExternalModule(node) || some(result.statements, isExternalModuleIndicator)) {
+            if (!isExternalModule(node) || getEmitModuleKind(compilerOptions) === ModuleKind.Preserve || some(result.statements, isExternalModuleIndicator)) {
                 return result;
             }
             return factory.updateSourceFile(
@@ -145,6 +146,9 @@ export function transformECMAScriptModule(context: TransformationContext): (x: S
         const args: Expression[] = [];
         if (moduleName) {
             args.push(moduleName);
+        }
+        if (getEmitModuleKind(compilerOptions) === ModuleKind.Preserve) {
+            return factory.createCallExpression(factory.createIdentifier("require"), /*typeArguments*/ undefined, args);
         }
 
         if (!importRequireStatements) {
@@ -238,9 +242,27 @@ export function transformECMAScriptModule(context: TransformationContext): (x: S
         return statements;
     }
 
-    function visitExportAssignment(node: ExportAssignment): VisitResult<ExportAssignment | undefined> {
-        // Elide `export=` as it is not legal with --module ES6
-        return node.isExportEquals ? undefined : node;
+    function visitExportAssignment(node: ExportAssignment): VisitResult<ExportAssignment | ExpressionStatement | undefined> {
+        if (node.isExportEquals) {
+            if (getEmitModuleKind(compilerOptions) === ModuleKind.Preserve) {
+                const statement = setOriginalNode(
+                    factory.createExpressionStatement(
+                        factory.createAssignment(
+                            factory.createPropertyAccessExpression(
+                                factory.createIdentifier("module"),
+                                "exports",
+                            ),
+                            node.expression,
+                        ),
+                    ),
+                    node,
+                );
+                return statement;
+            }
+            // Elide `export=` as it is not legal with --module ES6
+            return undefined;
+        }
+        return node;
     }
 
     function visitExportDeclaration(node: ExportDeclaration) {
