@@ -25,5 +25,5 @@ CompilerOptions::
   "configFilePath": "tsconfig.json"
 }
 Errors::
-[91merror[0m[90m TS6046: [0mArgument for '--module' option must be: 'none', 'commonjs', 'amd', 'system', 'umd', 'es6', 'es2015', 'es2020', 'es2022', 'esnext', 'node16', 'nodenext'.
+[91merror[0m[90m TS6046: [0mArgument for '--module' option must be: 'none', 'commonjs', 'amd', 'system', 'umd', 'es6', 'es2015', 'es2020', 'es2022', 'esnext', 'node16', 'nodenext', 'preserve'.
 
